css fixes from rebase changes
