+++ gateway/sds_gateway/templates/users/partials/captures_table.html
@@ -1,5 +1,5 @@
 <div class="table-responsive">
-    <table class="table table-striped table-hover" id="captures-table">
+    <table class="table table-hover" id="captures-table">
         <thead>
             <tr>
                 <th class="w-5"></th>
@@ -37,5 +37,17 @@
 
     .w-25 {
         width: 25%;
+    }
+
+    /* Ensure table-warning works consistently */
+    #captures-table tbody tr.table-warning {
+        --bs-table-bg: var(--bs-warning-bg-subtle);
+        background-color: var(--bs-warning-bg-subtle) !important;
+    }
+
+    /* add a max height to the table */
+    #captures-table {
+        max-height: 85%;
+        overflow-y: auto;
     }
 </style>
