+++ gateway/sds_gateway/templates/users/partials/review_create_dataset.html
@@ -1,12 +1,12 @@
 {% load custom_filters %}
 
 <!-- Dataset Information -->
-<div class="col-3 mb-4">
+<div class="col-4 mb-4">
     <div class="card">
         <div class="card-header bg-light">
             <h5 class="mb-0">Dataset Information</h5>
         </div>
-        <div class="card-body">
+        <div class="card-body" id="dataset-info-review-body">
             <div class="mb-3">
                 <label class="fw-bold">Dataset Name</label>
                 <p class="mb-0 dataset-name"></p>
@@ -22,7 +22,7 @@
         </div>
     </div>
 </div>
-<div class="col-7 mb-4">
+<div class="col-8 mb-4">
     <!-- Selected Captures -->
     <div class="card mb-4">
         <div class="card-header bg-light d-flex justify-content-between align-items-center">
